Adicionando endpoint que retorna nome do usuario logado
