+++ application/src/main/java/br/com/ucb/book/application/controller/UsuarioController.java
@@ -1,36 +1,31 @@
 package br.com.ucb.book.application.controller;
 
-import br.com.ucb.book.application.dto.request.LoginRequest;
-import br.com.ucb.book.application.dto.request.UsuarioRequest;
-import br.com.ucb.book.application.dto.response.TokenResponse;
+import br.com.ucb.book.application.dto.response.UsuarioResponse;
 import br.com.ucb.book.application.mapper.UsuarioDtoMapper;
-import br.com.ucb.book.domain.model.Usuario;
 import br.com.ucb.book.domain.service.UsuarioService;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/permitted")
 @RequiredArgsConstructor
+@RequestMapping("/api/usuario")
 public class UsuarioController {
 
     private final UsuarioService usuarioService;
 
     private final UsuarioDtoMapper usuarioDtoMapper;
 
-    @PostMapping("/cadastro")
-    @ResponseStatus(HttpStatus.CREATED)
-    public void cadastrar(@Valid @RequestBody UsuarioRequest usuarioRequest) {
-        Usuario usuario = usuarioDtoMapper.toModel(usuarioRequest);
-        usuarioService.criar(usuario);
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
+    public UsuarioResponse getUsuarioByToken(Authentication authentication) {
+        Jwt token = (Jwt) authentication.getPrincipal();
+        return usuarioDtoMapper.toResponse(usuarioService.findById(token.getClaim("id")));
     }
 
-    @PostMapping("/login")
-    @ResponseStatus(HttpStatus.OK)
-    public TokenResponse login(@Valid @RequestBody LoginRequest loginRequest) {
-        Usuario usuario = usuarioDtoMapper.toModel(loginRequest);
-        return usuarioDtoMapper.toTokenResponse(usuarioService.login(usuario));
-    }
 }
