+++ application/src/main/java/br/com/ucb/book/application/dto/response/UsuarioResponse.java
@@ -1,0 +1,11 @@
+package br.com.ucb.book.application.dto.response;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class UsuarioResponse {
+
+    private String nome;
+}
