+++ domain/src/main/java/br/com/ucb/book/domain/persistence/UsuarioPersistence.java
@@ -6,4 +6,6 @@ public interface UsuarioPersistence {
 
     Usuario criar(Usuario usuario);
 
+    Usuario findById(Long id);
+
 }
